Attempt at fixing empty nginx config bug
If nixy loses contact with Marathon it will overwrite the
nginx config with an empty configuration. This will
effectively take down all apps. In this situation it would
be more prudent to not touch the nginx config. This change
causes fetchApps() to return an error if no apps and no tasks
 where found.